+++ CHANGELOG.md
@@ -11,6 +11,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * First release to be published to [crates.io](https://crates.io/).
 
-[unreleased]: https://github.com/0xdea/{{project-name}}/compare/v0.1.0...HEAD
+[unreleased]: https://github.com/0xdea/haruspex/compare/v0.1.0...HEAD
 
 [0.1.0]: https://github.com/0xdea/haruspex/releases/tag/v0.1.0