+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.1.1] - 2024-11-29
 
 ### Added
 
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 * Change output directory's extension from `hpx` to `dec`.
+* Update doc workflow to include dependencies.
 
 ### Fixed
 
@@ -23,6 +24,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * First release to be published to [crates.io](https://crates.io/).
 
-[unreleased]: https://github.com/0xdea/haruspex/compare/v0.1.0...HEAD
+[unreleased]: https://github.com/0xdea/haruspex/compare/v0.1.1...HEAD
+
+[0.1.1]: https://github.com/0xdea/haruspex/compare/v0.1.0...v0.1.1
 
 [0.1.0]: https://github.com/0xdea/haruspex/releases/tag/v0.1.0